OurborosTag: don't take the current slot when comparing candidates
The methods `preferCandidate` and `compareCandidates` took a `SlotNo`,
corresponding to the current slot according to the wall clock. This `SlotNo`
was used to clip chains so that they no longer extend in the future. In the
future, this will become the responsibility of the chain sync client, so we
remove the argument.

Since the `Maybe (Chain b)` return type was used to return the clipped chain,
this is no longer necessary and the return type is now just `Bool`.

Another reason for doing this is that the `BlockFetchConsensusInterface`
record will need to use these two methods (as pure functions), without having
access to the current slot.

Remove the Ouroboros.Consensus.Protocol.Genesis module for now, we will
revisit it in the future.